+++ package.json
@@ -41,6 +41,7 @@
     "axios": "^1.6.7",
     "cookie-parser": "1.4.5",
     "eslint-plugin-check-file": "^2.7.1",
+    "eslint-plugin-unused-imports": "^3.1.0",
     "express": "4.18.2",
     "govuk_frontend_toolkit": "^9.0.1",
     "govuk-elements-sass": "^3.1.3",
