+++ src/routers/handlers/individual-psc-list/individualPscList.ts
@@ -1,12 +1,12 @@
 import { Request, Response } from "express";
+import { PrefixedUrls } from "../../../constants";
+import { logger } from "../../../lib/Logger";
+import { getLocaleInfo, getLocalesService, selectLang } from "../../../utils/localise";
 import {
     BaseViewData,
     GenericHandler,
     ViewModel
 } from "../generic";
-import { logger } from "../../../lib/Logger";
-import { PrefixedUrls } from "../../../constants";
-import { selectLang, getLocalesService, getLocaleInfo } from "../../../utils/localise";
 
 export class IndividualPscListHandler extends GenericHandler<BaseViewData> {
