+++ src/routers/handlers/individual-statement/individualStatement.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from "express";
-import { BaseViewData, GenericHandler, ViewModel } from "../generic";
-import { logger } from "../../../lib/Logger";
 import { PrefixedUrls } from "../../../constants";
-import { selectLang, getLocalesService, getLocaleInfo } from "../../../utils/localise";
+import { logger } from "../../../lib/Logger";
+import { getLocaleInfo, getLocalesService, selectLang } from "../../../utils/localise";
+import { BaseViewData, GenericHandler, ViewModel } from "../generic";
 
 interface IndividualStatementViewData extends BaseViewData { }
