+++ src/routers/individualPscListRouter.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response, Router } from "express";
-import { IndividualPscListHandler } from "./handlers/individualPscList/individualPscList";
 import { handleExceptions } from "../utils/async.handler";
+import { IndividualPscListHandler } from "./handlers/individual-psc-list/individualPscList";
 const router: Router = Router();
 
 router.get("/", handleExceptions(async (req: Request, res: Response, _next: NextFunction) => {
