+++ src/routers/rlePscListRouter.ts
@@ -1,7 +1,6 @@
-import { NextFunction, Request, Response, Router } from "express";
-import { RleListHandler } from "./handlers/rlePscList/rlePscList";
+import { Request, Response, Router } from "express";
 import { handleExceptions } from "../utils/async.handler";
-import { logger } from "../lib/Logger";
+import { RleListHandler } from "./handlers/rle-psc-list/rlePscList";
 const router: Router = Router();
 
 router.get("/", handleExceptions(async (req: Request, res: Response) => {
