+++ src/server.ts
@@ -1,8 +1,8 @@
 import fs from "fs";
 import http from "http";
 import https from "https";
-import { logger } from "./lib/Logger";
 import app from "./app";
+import { logger } from "./lib/Logger";
 
 // start the HTTP server
 const httpServer = http.createServer(app);
