+++ src/services/internal/pscVerificationService.ts
@@ -1,11 +1,11 @@
-import { Request } from "express";
 import { Resource } from "@companieshouse/api-sdk-node";
 import ApiClient from "@companieshouse/api-sdk-node/dist/client";
-import { Transaction } from "@companieshouse/api-sdk-node/dist/services/transaction/types";
-import { ApiErrorResponse } from "@companieshouse/api-sdk-node/dist/services/resource";
-import { createOAuthApiClient } from "../external/apiClientService";
 import { PscVerification, PscVerificationResource } from "@companieshouse/api-sdk-node/dist/services/psc-verification-link/types";
+import { ApiErrorResponse } from "@companieshouse/api-sdk-node/dist/services/resource";
+import { Transaction } from "@companieshouse/api-sdk-node/dist/services/transaction/types";
+import { Request } from "express";
 import { createAndLogError, logger } from "../../lib/Logger";
+import { createOAuthApiClient } from "../external/apiClientService";
 
 export const createPscVerification = async (request: Request, transaction: Transaction, pscVerification: PscVerification): Promise<PscVerificationResource> => {
     const oAuthApiClient: ApiClient = createOAuthApiClient(request.session);
