--- src/types/JsonData.ts
@@ -1,12 +1,0 @@
-type JSONValue =
-    | string
-    | number
-    | boolean
-    | JSONObject
-    | JSONArray;
-
-interface JSONArray extends Array<JSONValue> { }
-
-export interface JSONObject {
-    [x: string]: JSONValue;
-}
