+++ src/utils/session.ts
@@ -1,8 +1,8 @@
 import { Session } from "@companieshouse/node-session-handler";
+import { AccessTokenKeys } from "@companieshouse/node-session-handler/lib/session/keys/AccessTokenKeys";
 import { SessionKey } from "@companieshouse/node-session-handler/lib/session/keys/SessionKey";
 import { SignInInfoKeys } from "@companieshouse/node-session-handler/lib/session/keys/SignInInfoKeys";
 import { ISignInInfo } from "@companieshouse/node-session-handler/lib/session/model/SessionInterfaces";
-import { AccessTokenKeys } from "@companieshouse/node-session-handler/lib/session/keys/AccessTokenKeys";
 
 export function getSignInInfo (session: Session): ISignInInfo | undefined {
     return session?.data?.[SessionKey.SignInInfo];
