+++ test/mocks/all.middleware.mock.ts
@@ -1,5 +1,5 @@
-import mockSessionMiddleware from "./sessionMiddlewareMock";
 import mockAuthenticationMiddleware from "./authenticationMiddlewareMock";
+import mockSessionMiddleware from "./sessionMiddlewareMock";
 
 export default {
     mockSessionMiddleware,
