+++ test/routers/healthcheck.unit.ts
@@ -1,7 +1,3 @@
-import app from "../../src/app";
-import request from "supertest";
-import { PrefixedUrls } from "../../src/constants";
-
 describe("healthcheck tests", () => {
     it.skip("should show status 200", async () => {
